+++ CHANGELOG.md
@@ -3,3 +3,11 @@
 ## [1.01.01] — 2026-06-10 21:31
 - **✨ Enhancement**: bumped from `1.0.02` → `1.01.01`
 - Added + New Project button to extension popup
+
+## [1.02.01] — 2026-06-11 14:14
+- **✨ Enhancement**: bumped from `1.01.01` → `1.02.01`
+- Implement window screenshot cropping, standardized high-opacity click action circle, and nearest-button click description recognition
+
+## [1.03.01] — 2026-06-11 14:14
+- **✨ Enhancement**: bumped from `1.02.01` → `1.03.01`
+- Implement window screenshot cropping, standardized high-opacity click action circle, and nearest-button click description recognition